Rework CI/CD
Fetch full history when cloning repo.

Add --pre flag to pip install in order to allow installing
scikit-decide pre-releases.

Remove build-docs job, it will be performed in another workflow.

Modify build in order to let runners cache compilation of
chuffed, gecode and libminizinc.
An environment variable SKDECIDE_SKIP_DEPS can be set to skip
these compilations.  If unset, they are built as before.

It is better to use a dedicated environment variable as is done
here instead of an existing $GITHUB_* variable, this allows
anyone building skdecide in their CI to decide whether or not to
build these modules.

Modify Linux builds so that wheels for Python 3.7 and 3.8 are
built in distinct jobs, as with Mac or Windows.

Install cmake with pip on Linux docker runners, its compilation
takes time.