Do not force a specific clang-format version
Brew does not provide clang-format v9, so MacOS users may not be able
to easily run pre-commit locally.  We may try with v8, but since it is
unclear whether we really need a versioned dependency on clang-format,
let us try with an unversioned one for now, and we will get back to a
versioned dependency if discrepancies are found.