Use docker modularly
An initial docker image is now created for additional dependencies
through the new Dockerfile_x86_64_dev, build-skdecide_dev.sh and
build-pip-install.sh.
This image is cached to gain some time.

The skdecide_x86_64 image (building linux wheels) is created on top of
skdecide_dev through Dockerfile_x86_64, build-manylinux-wheels and
build-pip-install.sh.

The resulting docker image now contains all the files it created or loaded
and not just the output.

Fetching a file from the docker image can be done through a single
docker run --rm -v command in the .yml files

Also remove the setup of docker BUILD_KIT that wasn't used

build-pip-install.sh ensure that a build can still be done
manually as specified in the comments of build-manylinux.sh.
It is called both in build-skdecide_dev.sh, to be able to cache the
packages with the first docker image and in build-manylinux.sh
for manual builds.