Use PPO as RL solver in gym nb and reduce training time
This fix a "bug": the text was talking about PPO but code actually used
SAC instead.
The training time has also been reduced since this tutorial is meant to
run in a short time. A comment has been added to explain to the user how
to increase training time if he want to test.