Create specific binder env for each scikit-decide release
Summary:
During release.yml workflow, we push a new tag  defining the
corresponding binder environment, then build and cache it
on mybinder.org. Finally the binder links in the release doc point to
this specific environment.

Hypotheses:
- a branch named "binder" exists on origin with environment.yml file
  used to define a binder env
- the branch is up-to-date with necessary dependencies to make work the
  notebooks on master with master version of scikit-decide
- environment.yml file contains the dependency in scikit-decide, and it
  follows the format `scikit-decide[all]==...`

Process:
- we checkout binder branch
- we replace the spec for scikit-decide by the scikit-decide version
  corresponding to the release triggering the workflow
- we commit and push a tag named "binder_vx.x.x" where x.x.x stands for
  the inferred version number
- we force a build of the environment on mybinder.org to accelerate
  later call to it
- we use this binder env for links in the doc for the release

Note:  used a quite verbose way to define the BINDER_TAG variable.
This is to make it clear that we create it exactly the same way in both jobs
"create-binder-env" and "build-docs" , as there is no simple way to
share variables between jobs (without using artifacts).