DEBUG: Remove State.tasks_remaining
This set is heavy and makes State.copy() slow.

State contains the following attributes (among others):
* task_ids: a list of all task ids in the scheduling domain
* tasks_remaining: a set containing the ids of tasks still to be started
* tasks_complete: a set containing the ids of tasks that have been
      completed
* tasks_ongoing: a set containing the ids of tasks started and not
      paused and still to be completed
* tasks_paused: a set containing the ids of tasks that have been started
      and paused but not resumed yet
* tasks_progress: a dictionary where the key is a task id (int) and
      the value the progress of the task between 0 and 1 (float)
Remark: keys(tasks_progress) = tasks_ongoing + tasks_paused

The idea behind tasks_remaining is that some tasks may require
conditions which are not fulfilled, so tasks_remaining is a subset of
task_ids which contains tasks which can be started.

But at the beginning of simulation, this set may be large, and is copied
into all State instances.  We replace this member by
* tasks_unsatisfiable: a set containing the ids of tasks for which
      canditions are not fulfilled

We can use tasks_unsatisfiable to compute tasks_remaining by:

   task_ids - tasks_complete - tasks_progress - tasks_unsatisfiable

Copy of tasks_unsatisfiable will be much faster than tasks_remaining,
but the extra computations may cancel this gain, the overall speedup
(if any) must be confirmed on real cases.