Force a rebuild of binder environment at each nightly build
This is done by pushing a "dummy" commit to binder branch.
The commit only updates a comment with the nightly build name.
The comment itself does nothing, but the fact that the binder branch is
updated will trigger a new build the next time the binder environment
will be called.