Fix #133 (Docs is not building with node 16)
Call the function updateLinks() which references document (ie DOM)
inside a `mounted` hook.

Indeed the previous code was working with `yarn vuepress dev docs` but
not with `yarn vuepress build docs` because when *building*, vue uses a
server to generate the static website and usually `document` is only
available on client side.

However this is possible when inside the mounted hook
(https://vuepress.vuejs.org/guide/using-vue.html#browser-api-access-restrictions)

NB: no idea why this was working with node 14 though...