Make update of scikit-decide dependency more robust
In commit 4c4f00452, environment.yml is modified to depend on the exact
scikit-decide version being built.
The sed command assumed that environment.yml includes a line
  scikit-decide[all]==<some-version-number>
This commit makes the sed command more robust, it will now work whatever
extras section is used (it may also not be set).