Squash master doc commits on gh-pages between releases
The goal is to avoid a long unreadable chain of "publish documentation" commits, due
to the upload of master doc each night after a push on master.

This is done by:
- specifying the release concerned in commit message when deploying the
  release doc
- squashing the new commit created by the deploy doc action with the
  previous one if it has the same commit message first line (i.e.
  "publish documentation for master").
=> the first master doc commit after a release will not be merged with
the release doc commit, and we get an arborescence in gh-pages like

"publish documentation for master" > "publish documentation for 0.9.4" >
"publish documentation for master" > "publish documentation for 0.10.0"
> "publish documentation for master" > ...

Notes:
- as references like HEAD~1 sometimes does not work afer using the action
github-pages-deploy-action, we checkout again the repo before squashing
the commits.
- the doc deployment was done oint 2 places in build.yml, so we keep
  only one.