Fix minizinc configuration
When importing skdecide.hub, the path to minizinc binaries and to
solvers definition is updated with embedded versions of minizinc.

This worked great with install from releases but was not working when
using poetry, because paths are different.
Moreover, more folders than needed were potentially added to PATH
variable.

For release install, it is actually sufficient to look at directory
containing skdecide.hub.__init__.py as it should contain the solvers
configuration files *.msc and the bin/ directory with minizinc (and
the C++ library __skdecide_hub_cpp)

For poetry, except for gecode.msc, the solver configurations and the bin/
directory with minizinc are in the build directory generated by `poetry install`,
together with the C++ library  __skdecide_hub_cpp, which is used to find the proper
directory.
=> skdecide/hub/__init__.py is modified consequently.

For this to work, gecode.msc needs also to be copied to the directory in build/
containing the other solvers configurations like chuffed.msc. Indeed, gecode.msc
"mznlib" field refers to a path relative to it which actually exists in the latter directory.
=> a postbuild_dev.py script is added doing this and the documentation
for developpers is amended to show how to use it.