Avoid installing prereleased dependencies in github action
The way the scikit-decide library was installed for tests was designed
so that pip find scikit-decide (in prelease mode) in the "wheels" directory.
However a side effect was that it also allowed pip to install preleased
dependencies, which can lead to some bugs, see issue #192.

To avoid this we rather specify explicitely the path to the wheel (but
unsing wildcards to get easily the exact name) and get rid of the
"--pre" option.

We do that in a separate step for better readibility.

Fix #192.