Replace try..except by an method existence check to ease debugging
The call to init_model() was inside a try.. except block because some
solver does not have such method.
We replace this block by a test wether the method exists so that an
error inside the init_model() will be seen to better understand why a
solve does not work as expected.