Install last nightly build in binder env for master version
Here we set manually the url of the nightly build.
The github action will be modified so that
- a new nightly build will update this url
- the environment for a release will empty this url

Note that we do not touch the environemnt.yml so that previous release
of scikit-decide should already been installed with all its
dependencies in docker cache.
It should speed-up the new image build as only updated dependencies
should be downloaded.