Fix issue when building doc on node 18
The ssl behaviour change in last node version. This breaks the doc build
with the current dependencies. We can force to use legacy openssl
provider. It could be a security issue but not here as there is no
interaction with external network during the build process.

Another solution would be to find the modification in the dependencies that removes this
error.