Preinstall gym==0.21.0 in github workflows
Indeed, new versions of pip/setuptools enforce pep 517 which is not
respected by gym 0.21.0 install. So we preinstall with option
--no-use-pep517 so that the installation of scikit-decide[all] works.

Note that it could be good to upgrade to gymnasium 0.26 and later but
this is not yet possible because of a conflict between recent versions
of ray[rllib] and stable-baselines3 which are each fixing a different
version of gymnasium.