Change way action space is discretized  (#205)
This commit change the way a Box action is discretized.
It is inspired from the gym.spaces.Box.sample method. For unbounded spaces it builds on the quantile distribution of the Cauchy distribution (using a scale of 1).
An alternative could be derived from the logit function: log(1+p) - log(1-p) for example. But the current implementation has a larger spread.