Add --all-files option to precommit command to avoid skipping the test
No tests were actually done by the linter action because `pre-commit
run` look only files in the index by default.
When checkout a whole PR, no files are in the index so no files are
checked.

Previously, the job returned something like

	Trim Trailing Whitespace.............................(no files to check)Skipped
	Fix End of Files.....................................(no files to check)Skipped
	Check Yaml...........................................(no files to check)Skipped
    	Check for added large files..........................(no files to check)Skipped
    	Pretty format JSON...................................(no files to check)Skipped
    	isort (python).......................................(no files to check)Skipped
    	black................................................(no files to check)Skipped
    	nbqa-isort...........................................(no files to check)Skipped
    	nbqa-black...........................................(no files to check)Skipped
    	nbqa-pycln...........................................(no files to check)Skipped
    	nbstripout...........................................(no files to check)Skipped

Now the checks are done (and should read as "Passed")
We make a `pre-commit run --all-files` in this commit to have the job
pass.

We also remove the pretty-format-json hook which was tampering with the
notebooks (oddly only remotely on the github runner, not locally on my
laptop), and was not very useful for the repository anyway.