Fix a serialization issue by putting NewType declarations outside of class definition
We faced serialization issues with ray.rllib solver and python 3.10.
This seemed to be due to NewType definition inside Domain class.
Actually the NewType assumed that the resulting is belonged by the
module and not the class which causes pickle issues.

See for instance https://stackoverflow.com/a/4677063 for similar issue.