Update python allowed
We start now from 3.8 and end with 3.11.
We build a wheel for 3.11.

We also specify a scipy version for which a wheel exists from python 3.8
to python 3.11. It is simplifying the way poetry resolves dependencies
for the dev environment.

To generate the manylinux wheel, we need to ensure that the latest
of manylinux docker image is used so that python3.11 exists on it.