Update notebooks for new release with the proper version of scikit-decide
When tagging a new release vx.y.z, we create an associated tag
notebooks-vx.y.z with notebooks installing (on colab) the appropriate
release version instead of the nightly wheel.

Then the doc for the release will point towards those notebooks so that
the examples use the release.

If on fork, we install the release from Test.Pypi where the release
should have been deployed.