Implement callbacks for python solvers
We add a callback hook for:
- ARS
- CGP
- lazy A*
- LRT A*
- MaxentIRL
- POMCP
- sb3 wrapper

A callback is a function, called at each solver's iteration,  taking the solver as argument
and returning a boolean. If returning True, the solver stops.

We also make some intermediate results attributes from the solver to be
accessible in the callback.

And add limited docstring so that the signature of solvers __init__()
will show in the doc with the callback behaviour documented.