Ensure compatibility with openap 2.0
The fuel flow model has changed and FuelFlow.enroute() take vs (vertical
speed) as argument instead of path_angle.

So we adapt AircraftPerformanceModel.compute_fuel_consumption() to take
also vs as argument.
To be compatible with previous version, we make the conversion (vs->
path_angle) in case we still have path_angle in signature (openap
<=1.5).

Note that we needed to change expected values in `test_perf_model()`:
- for openap model, as they improves (and thus changed) the computation.
  We have 2 sets of value according to whether we are <2.0 or >=2.0
- for poll-schumann model as it was not correct before:

	altitude_next = altitude_current + speed_current * delta_time * math.radians(path_angle)

  should already have been

	altitude_next = altitude_current + speed_current * delta_time * math.tan(math.radians(path_angle)) / ft

  (with ft the constant to pass from feet to meters).

Now the test evaluates True both with openap 1.5 and 2.0.