Fix return hint for EnumerableSpace.get_elements() and simplify implementations
In code, the result of `EnumerableSpace.get_elements()` is used for:
- `for a in elements`
- `if a in elements`
- `elements[i]`
- `len(elements)`
- `random.sample(elements, 2)`

So this is actually expected to be `Sequence`.
We fix the annotation and simplify the implementation by removing the
wrapping in numpy.array() as it is not expected and even sometimes
re-wrap in a numpy.array() in the code.