Fix autodoc
The new solver characteristic Maskable has a (hidden) attribute initialized to None.
Thus the method  `get_ref()` does not work on it, and updating the
types_dict for the code generator was failing. As this attribute is not
used in type annotations of methods (which was the purpose of
types_dict) this is not a real issue. So we merely skip adding a type to
it when member["ref"] does not exist.