Implement GNN based PPO graph->node for ray.rllib framework
The custom model follows what has been done for sb3 framework:
- we predict values as before with GNN feature extraction + reduction to
  a fixed number of features + MLP
- we predict actions with a single GNN which predict logits for each
  node without knowing in advance their number.

To use that feature, we need to set the attribute graph_node_action to
True. The doc of RayRLlib has been enriched to explain what is happening
in that case.

The main issue here is that we have also to pad rollout buffers entries
corresponding to action_dist_inputs (the logits) as the number of
available actions is potentially varying from an observation to another.