Implement a scikit-decide PDDL domain based on plado library (#474)
* Implement a scikit-decide PDDL domain based on plado library

Plado is an alternative to UP for simulating PDDL problems, also simulating Probabilistic PDDL problems, even though without its own planners/solvers:
https://github.com/massle/plado/tree/main

It is in particular way faster to list applicable actions in a given
state (and does not need to enumerate all actions).

This code is inspired by
https://github.com/massle/plado/blob/60958c34105c01ec43f0dae8247dae889272220a/examples/skdecide_domain.py

To enable RL solvers use, we add a encoding of state and action into
simple gym spaces. This will be tractable only for small problems
though.

In later work, we intend to
- encode states as graphs (e.g. to use GNNs)
- keep actions represented as tuples to predict them in a autoregressive way:
  action_type -> param0 -> param1 ...

* Test ppddl plado domain