Implement autoregressive PPO for sb3 applied to plado PDDL/PPDDL domains (#478)
* Move base sb3 classes for buffers and algos to be used by several customized ones

To be used by GNN-based algos and later by algos for composite actions with
autoregressive prediction of components.

* Implement a masked version of MultiDiscreteSpace

This will be useful to model variable-length multidiscrete actions,
e.g. parameteric actions whose arity depends on its first component,
the action type (for PDDL domains).

An element of this space can have -1 components corresponding to
non-existing components. Otherwise it is still between 0 and nvec[i]-1.

* Implement state_sample in plado domain

As transition proba and cost are computed together by plado engine we
override _state_sample(), and update get_transition_value and
get_next_state/get_next_state_transition that may be directly called by
algos (e.g.lazy-A*).

get_next_state cache the last value computed and usually
get_transition_value is called afterwards. But if the value is not
cached anymore, we call again the plado engine.

* update plado domain for autoregressive action prediction

We add a multidiscrete encoding of the actions to be used by our
autoregressive sb3 algo.

Beware that this is not a regular MultiDiscrete Space as we allow -1
values, indicating that the component is actually not needed (for
actions with fewer parameters than others)

* Implement autoregressive sb3 PPO for PDDL domains