Add possibility to have negated atoms in goal and action precondition
We add a node feature encoding the negation of an atom.
The scheme predicate node will have this feature to 1 if it is encoding
a precondition or a goal with a negated atom.

The edge label ADD and DEL are merged into a single label EFFECT as the
add/del part is encoded via the "negated atom" node feature.