Implement a custom GNN class
As thg.nn.models.basic_gnn.BasicGNN, it wraps MessagePassing layers.
Main differences:
- possibility to add an encoding/decoding layer as first/last layer (a
  plain MLP on nodes features)
- possibility to repeat a single convolutionnal layer instead of
  duplicate it (less params to learn)

We add tests to showcase its use with autoregressive graph PPO
algorithm.

To make it easier to the user, we update GraphFeaturesExtractor and
Graph2NodeLayer so that it automatically fill the arg "in_channels" (and
"out_channels" for the latter) from the observation space dimension.