Encode plado state as a Lifted Learning Graph  (#482)
* Encode plado state as a Lifted Learning Graph

We add a state_encoding to plado domains (mainly) corresponding to the
LLG presented in

    Chen, D. Z., Thiébaux, S., & Trevizan, F. (2024).
    Learning Domain-Independent Heuristics for Grounded and Lifted Planning.
    Proceedings of the AAAI Conference on Artificial Intelligence, 38(18), 20078-20086.
    https://doi.org/10.1609/aaai.v38i18.29986

Hypotheses (for now):
- no fluents
- no negative atoms in preconditions or goals
- no conditional/probabilistic effects

Main diff with the paper:
- we add a node feature to specify that a predicate is static
- we use a simpler index function (one hot encoding) as we simply intend
  to be problem-independent (and not domain-independent) so we know the
  arity of each predicate in advance

* Add possibility to have negated atoms in goal and action precondition

We add a node feature encoding the negation of an atom.
The scheme predicate node will have this feature to 1 if it is encoding
a precondition or a goal with a negated atom.

The edge label ADD and DEL are merged into a single label EFFECT as the
add/del part is encoded via the "negated atom" node feature.

* Add fluents in LLG representation

* Test pddl domain with llg encoding solved by sb3 autoregressive gnn