Allow to load a autoregressive graph2node sb3 algo with different action space
When ome action components are graph nodes the algo can apply on a
different action space. Still
- len(action_space.nvec) ie nb of action components cannot change
  (as the model has a network by component)
- independent components must keep the same dim (as it is hard coded in
  the network in that case)

This can be useful when training a autoregressive GNN model on a PDDL
domain and then apply on a similar pddl domain but different problem
(with same actions but different objects)

To achieve that we just need to be careful to update the algo
action_space. The parameters of the torch layers are still loaded as
before.