Allow to load a sb3 autoregressive graph ppo on different action space (#497)
* Allow an observation formatter in rollout utils

Avoid displaying lengthy initial observation, in particular for graph
instances, and use instead a more friendy user formatter.

* Update StableBaseline.load() to take into account new algo params

We allow now to change learning_rate, n_steps, and so on when loading
a sb3 algo.

* Allow to load a autoregressive graph2node sb3 algo with different action space

When ome action components are graph nodes the algo can apply on a
different action space. Still
- len(action_space.nvec) ie nb of action components cannot change
  (as the model has a network by component)
- independent components must keep the same dim (as it is hard coded in
  the network in that case)

This can be useful when training a autoregressive GNN model on a PDDL
domain and then apply on a similar pddl domain but different problem
(with same actions but different objects)

To achieve that we just need to be careful to update the algo
action_space. The parameters of the torch layers are still loaded as
before.

* Update autoregressive gnn sb3 example on pddl with llg encoding

- train + rollout + save on a first problem (blocksworld with 3 blocks)
- load + train + rollout on another problem (with 4 blocks)
- change rollout formatters to avoid unreadable logs