Remove intermediate methods not supposed to be overriden by the user (#504)
The library has the convention that the public methods are autocastable
and have private counterparts (with leading "_") that are to be
overriden by the user.
But sometimes these counterparts had important code not to be lightly
override.

For instance:
- domain.reset() is autocastable and calls _reset()
- domain._reset() code is setting domain._memory, calling
  _state_reset(), and getting the corresponding observation
- domain._state_reset() is the method actually to be overriden by the
  user (not implemented by default at the higher level `Initializable`.

To simplify,
- the boilerplate code is now put directly in domain.reset()
- domain._reset() is removed to avoid confusion
- domain.state_reset() stands to be implemented

An advanced user can still override domain.reset() by being careful to
manage domain._memory and use @autocastable decorator.

We do the same for
-_step
- _sample
- _get_enabled_events
- _is_enabled_event
- _get_applicable_actions
- _render