Simplify llg encoding for plado domains
As we are not trying to be domain independent, it may be sufficient to
encode only the goal and state and drop the actions encoding from the
graph. This allow to drastically diminish the graph size.

e.g. for an example with blocksworld 3 blocks the initial state goes
from
nodes ~ (113, 20), edges ~ (346, 4)
to
nodes ~ (40, 7), edges ~ (136, 1)
and even, when dropping static facts, to
nodes ~ (23, 6), edges ~ (76, 1)

(nb of nodes/edges, nb of node features/edge features)

More precisely we introduce several options to LLgEncoder:
- encode_actions: boolean deciding to include the subgraph corresponding
  to actions or not (now False by default)
- simplify_encoding: drop node/edge features not needed by the domain.
  Once the initial graph encoded we drop the node feature that are not
  used except the ones that could be useful by the state itself.
- encode_static_facts: include or not static predicates (can be dropped
  if problem independance not necessary, or if not needed to
  distinguish between problem as for domain blocksworld)

And we add an option for PladoDomain:
- llg_encoder_kwargs: dict allowing to specify these new options

To adapt the sb3 autoregressive gnn ppo, we update the default
n_graph2node_components so that the dependent components starts with
the first one not having a None in
`action_components_node_flag_indices`.